requirements.txt
Flask==0.10.1
Flask_JWT==0.3.2
Flask_RESTful==0.3.5
orator==0.8.2
paho_mqtt==1.2
gunicorn
mysqlclient
Flask-Mail